0000138: add ability to zip selected files into a new archive add Zip functionality for files
0000168: ability to copy a file
add Copy functionality for files, including captions and tags

0000003: api hook to manage deletions
finish kfm_cmsHooks_allowedToDeleteFile() integration.

remove 'image_data' part of kfm_attributes data. also, remove kfm_showIcon function call. This has the effect of vastly improving the speed at which large directories of images are shown.
remove setThumbnail() from image() constructor. it is not needed at that point, and causes new directories to appear to be empty (takes a long time to generate the thumbnails)
remove inArray() function. turns out there is a kfm_inArray() function as well, so inArray() is not necessary

change file dragging so that the list of files can't cause scroll bars if you try to drag the list out of the window.

only show tooltips if not currently dragging files

change contextmenu links so that menu itself vanishes as soon as a link is pressed (instead of after the resulting function (a prompt, for instance) has been called).

fix deletion of multiple files so that an error is returned as soon as it happens (instead of trying every file in the array then returning nothing)