upgrade jquery
also, use minified version instead of packed. the unpacking of the packed version takes longer than the bandwidth needed to load the minified version.

http://www.julienlecomte.net/blog/2007/08/13/
"using Packer’s redundancy reduction algorithm adds a delay at the time the script gets executed (because the script needs to be unpacked and then eval’ed) On my laptop, this delay is about 200 msec"