the _(str,context) function will become the main translation function. (doesn't actually translate at the moment...)
str:     string to be translated.
context: context of the string (the plugin it's in, etc - differentiates between "rename" for directories and "rename" for files, for example)

str can also be an array of the form [str,context]

there will be another function __() for substitution strings such as "rename %1 to %2" - that's to allow a coder to choose the faster _() function when it is known that there won't be any substitution.