add a lazy loader for functions. based on an idea I noticed on the Ajaxian blog.
basically, all asynchronous functions that are likely not going to be called in any particular session are replaced with function stubs. when called, those stubs fetch the real js code from the server, replace  themselves with that js, then run.

this should be applied to synchronous functions (functions that return a value), or to functions that are guaranteed to be called at least once immediately upon loading KFM.