number of changes/improvements
add code necessary to send files back to CKeditor
the admin panel link's display is optional (default to yes)
add a restriction, letting the max image width/height to be set. default to 1024x768. this can be changed to a huge number if you don't want any restrictions.
there's also code in there to allow images to be tinted by HSL, using ImageMagick.
also, code is there to upload files into specified directories (not used by the GUI, but useful for 3rd-party apps that use KFM's functions).
a few other things are in there, but I can't remember them all!