optimisation of icon-view file display
separated list-view and icon-view into distinct functions which can be individually optimised.
created a directory with 1000 files in it. before starting, this loaded in 38472ms, 61522 function calls.
after optimisation, this loaded in 11122ms, and 47223 calls.

it kills Chrome, but that's a problem to be worked out tomorrow.
tested in FF 3.5, Opera 10, Konqueror, Chrome 5.0.307.9