first stab at an RPC file.
this file only includes a method for deleting a file. you need to know its ID first. also, authentication must be navigated first.

I'm using it in a CMS, where logging into the CMS also logs into KFM, and the delete function is called through AJAX, so the session ID is the browser's.